// prodis-prodiswcl/src/app/modules/prodisapi/model/partTime.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TipoRipropPt } from './tipoRipropPt';


export interface PartTime { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    id?: number;
    idProspettoProv?: number;
    nPartTime?: number;
    orarioSettContrattualeMin?: number;
    orarioSettPartTimeMin?: number;
    tipoRipropPt?: TipoRipropPt;
}